Tidy for normal manpage requirements/conventions
Also while we're in here, fix some formatting glitches, some things that
don't follow the normal manpage requirements or font conventions, drop
some formatting markup that isn't actually doing anything, and the note
that it was generated by docbook-to-man which is presumably responsible
for most or all of the above.

Updating the rest of the manpages similarly is left as an exercise for
people/days with more spare time.