Respect the configure --docdir override
Patch from RedDwarf, Closes ticket:1758 (for oggz).

There is more that could be done here, right now the install-data target
installs everything that doxygen generates into $docdir, which with the
current doxygen configuration includes both html and latex docs, the
latter of which aren't actually in a user-presentable form without doing
a further build of them to convert to the desired format.

Probably the html docs should be installed to htmldir, and the latex
docs should not be installed at all in their current form.  If people
do use the latter, then adding top level targets to fully generate
them wouldn't be the silliest thing to do, and then they should be
installed to dvidir/pdfdir/psdir as appropriate.

But so far, being able to alter $docdir is all that was asked for, and
this patch does exactly that.  The rest can wait for someone with the
right itch for now.